+++ ch12/ex12_32.cpp
@@ -1,38 +1,38 @@
 //
-//  ex12_27.cpp
-//  Exercise 12.27
+//  ex12_32.cpp
+//  Exercise 12.32
 //
-//  Created by pezy on 12/31/14.
-//  Copyright (c) 2014 pezy. All rights reserved.
+//  Created by pezy on 1/1/15.
+//  Copyright (c) 2015 pezy. All rights reserved.
 //
-//  The TextQuery and QueryResult classes use only capabilities that we have already covered.
-//  Without looking ahead, write your own versions of these classes.
+//  Rewrite the TextQuery and QueryResult classes to use a StrBlob
+//  instead of a vector<string> to hold the input file.
 
-#include "ex12_27_30.h"
-#include <sstream>
+#include "ex12_32.h"
+#include <sstream> 
 #include <algorithm>
 
-TextQuery::TextQuery(std::ifstream &ifs) : input(new vector<string>)
+TextQuery::TextQuery(std::ifstream &ifs) : input(new StrBlob)
 {
-    LineNo lineNo{0};
+    StrBlob::size_type lineNo{0};
     for (string line; std::getline(ifs, line); ++lineNo) {
         input->push_back(line);
         std::istringstream line_stream(line);
         for (string text, word; line_stream >> text; word.clear()) {
             // avoid read a word followed by punctuation(such as: word, )
             std::remove_copy_if(text.begin(), text.end(), std::back_inserter(word), ispunct);
-            // use reference avoid count of shared_ptr add.
+            // use reference avoid count of shared_ptr add. 
             auto &nos = result[word];
-            if (!nos) nos.reset(new std::set<LineNo>);
+            if (!nos) nos.reset(new std::set<StrBlob::size_type>);
             nos->insert(lineNo);
         }
     }
 }
 
-QueryResult TextQuery::query(const string& str) const
+QueryResult TextQuery::query(const string& str) const 
 {
     // use static just allocate once.
-    static shared_ptr<std::set<LineNo>> nodate(new std::set<LineNo>);
+    static shared_ptr<std::set<StrBlob::size_type>> nodate(new std::set<StrBlob::size_type>);
     auto found = result.find(str);
     if (found == result.end()) return QueryResult(str, nodate, input);
     else return QueryResult(str, found->second, input);
@@ -41,7 +41,9 @@ QueryResult TextQuery::query(const string& str) const
 std::ostream& print(std::ostream &out, const QueryResult& qr)
 {
     out << qr.word << " occurs " << qr.nos->size() << (qr.nos->size() > 1 ? " times" : " time") << std::endl;
-    for (auto i : *qr.nos)
-        out << "\t(line " << i+1 << ") " << qr.input->at(i) << std::endl;
+    for (auto i : *qr.nos) {
+        ConstStrBlobPtr p(*qr.input, i);
+        out << "\t(line " << i+1 << ") " << p.deref() << std::endl;
+    }
     return out;
 }
